+++ packages/nextjs/src/client-boundary/hooks.ts
@@ -19,6 +19,7 @@ export {
   isEmailLinkError,
   isKnownError,
   isMetamaskError,
+  isReverificationCancelledError,
   EmailLinkErrorCode,
   EmailLinkErrorCodeStatus,
 } from '@clerk/clerk-react/errors';